Dumps LSP server stdout and stdin logs on test failure

+++ tests/server/tests/conftest.py
@@ -3,7 +3,7 @@ import shutil
 import signal
 import sys
 from pathlib import Path
-from typing import Iterator
+from typing import Iterator, Optional
 
 import pytest
 
@@ -25,6 +25,9 @@ def client(request: pytest.FixtureRequest) -> Iterator[LFortranLspTestClient]:
 
     server_log_path = f"{request.node.name}-server.log"
     client_log_path = f"{request.node.name}-client.log"
+    stdout_log_path = f"{request.node.name}-stdout.log"
+    stdin_log_path = f"{request.node.name}-stdin.log"
+    gdb_log_path = f"{request.node.name}-gdb.log"
 
     config = {
         "LFortran": {
@@ -73,30 +76,44 @@ def client(request: pytest.FixtureRequest) -> Iterator[LFortranLspTestClient]:
         "--min-retry-sleep-time-ms", str(config["LFortran"]["retry"]["minSleepTimeMs"]),
         "--max-retry-sleep-time-ms", str(config["LFortran"]["retry"]["maxSleepTimeMs"]),
         "--extension-id", "lcompilers.lfortran",
+        "--execution-strategy", "concurrent",  #-> "parallel" or "concurrent"
     ]
 
-    client = LFortranLspTestClient(server_path, server_args, None, 1000, config, client_log_path)
+    def print_log(log_path: str, heading: str) -> None:
+        header = f"~~ {heading} [{log_path}] ~~"
+        border = "~" * len(header)
+        print(file=sys.stderr)
+        print(border, file=sys.stderr)
+        print(header, file=sys.stderr)
+        print(border, file=sys.stderr)
+        print(file=sys.stderr)
+        if os.path.exists(log_path):
+            with open(log_path) as f:
+                print(f.read(), file=sys.stderr)
+        else:
+            print(f"Log file does not exist: {log_path}")
 
     def print_logs() -> None:
-        if os.path.exists(client_log_path):
-            print()
-            print("~~~~~~~~~~~~~~~~~", file=sys.stderr)
-            print("~~ Client Logs ~~", file=sys.stderr)
-            print("~~~~~~~~~~~~~~~~~", file=sys.stderr)
-            print()
-            with open(client_log_path) as f:
-                print(f.read(), file=sys.stderr)
-        if os.path.exists(server_log_path):
-            print()
-            print("~~~~~~~~~~~~~~~~~", file=sys.stderr)
-            print("~~ Server Logs ~~", file=sys.stderr)
-            print("~~~~~~~~~~~~~~~~~", file=sys.stderr)
-            print()
-            with open(server_log_path) as f:
-                print(f.read(), file=sys.stderr)
+        print_log(client_log_path, "Client Logs")
+        print_log(stdin_log_path, "Standard Input")
+        print_log(stdout_log_path, "Standard Output")
+        print_log(server_log_path, "Server Logs")
+        print_log(gdb_log_path, "GNU Debugger (GDB)")
 
+    client: Optional[LFortranLspTestClient] = None
     logs_printed = False
     try:
+        client = LFortranLspTestClient(
+            server_path=server_path,
+            server_params=server_args,
+            workspace_path=None,
+            timeout_ms=1000,
+            config=config,
+            client_log_path=client_log_path,
+            stdout_log_path=stdout_log_path,
+            stdin_log_path=stdin_log_path
+        )
+
         with client.serve():
             # Steps abstracted by context manager:
             # 1. Send request: initialize
@@ -105,13 +122,18 @@ def client(request: pytest.FixtureRequest) -> Iterator[LFortranLspTestClient]:
             # 4. Send request: shutdown
             # 5. Send notification: exit
             yield client
-    except Exception as e:
-        print_logs()
-        logs_printed = True
+    except BaseException as e:
+        if not isinstance(e, SystemExit) or e.code != 0:
+            print_logs()
+            logs_printed = True
         raise e
     finally:
-        if hasattr(client, 'server') and not client.server.poll():
-            print('Server did not terminate cleanly, terminating it forcefully ...', file=sys.stderr)
+        if client is not None and hasattr(client, 'server') \
+           and not client.server.poll():
+            print(
+                'Server did not terminate cleanly, terminating it forcefully ...',
+                file=sys.stderr
+            )
             os.kill(client.server.pid, signal.SIGKILL)
             if not logs_printed:
                 print_logs()
